Refactored to an abstract test case for repositories.
